requirements.txt
openai
customtkinter==5.2.2
python-dotenv==1.0.0
playwright==1.40.0
